Gjort en god del refaktorering rundt domenemodeller for å harmonisere og dobbeltsjekke.. Er ikke helt i mål. ES-testene feiler plutselig uten grunn..

+++ src/main/kotlin/no/nav/klage/oppgave/api/view/OppgaveView.kt
@@ -9,7 +9,7 @@ const val YTELSE_SYK = "Sykepenger"
 const val YTELSE_FOR = "Foreldrepenger"
 
 data class OppgaverRespons(
-    val antallTreffTotalt: Int,
+    val antallTreffTotalt: Long,
     val oppgaver: List<Oppgave>
 )
